pyproject.toml
[project]
name = "dio-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Felipe Vieira",email = "felipedc.dev@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "databases[aiosqlite] (>=0.9.0,<0.10.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-mock (>=3.15.1,<4.0.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
